Prototype a uSDHC example
Format an SD (HC) card as a FAT16/32 filesystem with an MBR. Plug it
into your 1170EVK, flash the example, and run the example. If all goes
well, your board's LED will turn on. You can then remove the SD card and
use your host to view the secret file.

This depends on a WIP imxrt-usdhc driver. See the imxrt-usdhc README for
more information.

  https://github.com/mciantyre/imxrt-usdhc

This needs a review. It should also be ported to the Teensy 4.1.