Adds a SAI driver to the tree
A SAI driver lets a SaiTx and SaiRx be optionally constructed given pins.

A Sai[Tx/Rx] lets channels be constructed be constructed given a data pin. This is
a trait impl such that only available channels may actually be
constructed. In the case of SAI1 that means 4 channels. In the case of
other SAI instances only 1 channel.

In the end only each level manipulates certain registers to setup
channels appropriately so I believe safety can be kept.

Read/Write functionality can then be built on arrays of tx/rx channels
on top of private traits and impls per channel perhaps.

The end goal is pretty simple though, provide an interface for streaming
audio frames (sets of channel samples) in or out of a SAI data channel.