Add pll4 clock module and sai clock modules
Provides a pll4 reset function with all the options to set the
div_select, num, denom values with asserts ensuring bounds are kept
based on the data sheet. It would be nice to assert these at compile
time if possible and could be a future improvement to ensure the pll
options are valid.

The sai clock gating, selection, and divisor settings are also added
enabling the full clocking of a sai peripheral from the audio pll to the
peripheral input clock.