Integrate Channel<DMA_INST> with controller impls
Peripherals on all non-1180 MCUs remain compatible with all DMA
controllers. Although we haven't allocated the second DMA controller on
the 1170, my understanding is that all peripherals work with all
controllers. (If that's not right, we can now fix that.)

By implementing the dma::WorksWith trait, an 1180 peripheral fixes
itself to one of the two DMA controllers. I provided the implementation
for LPUART1 and eDMA3. I added the eDMA4 channels for testing
convenience (below), but they're otherwise unused.

This commit introduces type changes in the logging package. I'm
optimizing this refactor for the existing support; these changes won't
work if / when we extend the logging package to support the 1180.

If a user pairs an eDMA4 channel with LPUART1, here's the error the
compiler produces.

    error[E0308]: mismatched types
       --> examples/async_dma_uart.rs:30:30
        |
    30  |             console.dma_read(&mut channel, &mut buffer).await.unwrap();
        |                     -------- ^^^^^^^^^^^^ expected `3`, found `4`
        |                     |
        |                     arguments to this method are incorrect
        |
        = note: expected mutable reference `&mut imxrt_dma::channel::channel_impl::Channel<3>`
                   found mutable reference `&mut imxrt_dma::channel::channel_impl::Channel<4>`
    note: method defined here
       --> imxrt-rs/hal/src/chip/drivers/dma.rs:228:12
        |
    228 |     pub fn dma_read<'a, const DMA_INST: u8>(
        |            ^^^^^^^^

    error[E0308]: mismatched types
       --> examples/async_dma_uart.rs:33:31
        |
    33  |             console.dma_write(&mut channel, &buffer).await.unwrap();
        |                     --------- ^^^^^^^^^^^^ expected `3`, found `4`
        |                     |
        |                     arguments to this method are incorrect
        |
        = note: expected mutable reference `&mut imxrt_dma::channel::channel_impl::Channel<3>`
                   found mutable reference `&mut imxrt_dma::channel::channel_impl::Channel<4>`
    note: method defined here
       --> imxrt-rs/hal/src/chip/drivers/dma.rs:214:12
        |
    214 |     pub fn dma_write<'a, const DMA_INST: u8>(
        |            ^^^^^^^^^