Prefer defmt throughout examples
Most of the boards we support have a debugging interface for (defmt)
RTT. If we lean into defmt as our logging solution, we don't need
imxrt-log throughout our examples. This commit removes imxrt-log from
board and all the examples. This will make it easier for me to refactor
the BSP and our examples.

Since it removes logging, this commit handicaps folks who develop and
debug on a Teensy 4. I left a note in the Teensy 4 board module about
how we could add a defmt transport later.